Added skeleton of gRPC server
The service has not been added yet as we are creating a worker pool to demonstrate channels and workgroups to fan out/fan in requests in parallel threads. We are using a stop channel in this, which will be plugged into ctrl-c signal. We won't do auth so we won't add a shutdown command to the API, this will be for further more advanced server programming tutorial.