Fixed error type to have an Error method.
This means that now the generated protobuf error type can be used directly as an error return without having to define it twice.
The generated code should already have this in it, but doesn't. It has a map to forward and reverse access it from an iota constant but doesn't integrate either the Error method to implement the error interface, nor does it use the correct type on the values in the map.