Completed test from library version with service.
Some issues with the round robin job distribution counter... confusion about the difference between the *number* of workers versus the maximum index of the worker channel slice. Not sure if it should be written differently, trying to avoid adding a -1 in the loop as this is an unnecessary operation to do repeatedly on the worker count, so it's done when it's initialised.