Updated go install version and more notes re generator.
Go is now officially to 1.18, and generics are formally supported, and there is many small improvements to the runtime and compiler that are worth the upgrade.

Just some notes further clarifying about the prohibition of circular dependencies in Go, and the idea of creating a data specification separate from the implementation.